small fixes and docs updates

+++ src/commands/build.ts
@@ -34,6 +34,12 @@ export async function command({cwd, config}: DevOptions) {
   const dependencyImportMapLoc = path.join(config.installOptions.dest, 'import-map.json');
   const dependencyImportMap: ImportMap = require(dependencyImportMapLoc);
 
+  if (allRegisteredWorkers.filter(([id]) => id.startsWith('build')).length === 0) {
+    console.error(chalk.red(`No build scripts found, so nothing to build.`));
+    console.error(`See https://www.snowpack.dev/#build-scripts for help configuring your build.`);
+    return;
+  }
+
   rimraf.sync(finalDirectoryLoc);
   mkdirp.sync(finalDirectoryLoc);
   if (finalDirectoryLoc !== buildDirectoryLoc) {
